Remove git branch from CommandRunner's reponsibility
